move empty state into table
