+++ src/scenes/AcceleratorRouter/Cohorts/CohortListView.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { useHistory } from 'react-router-dom';
 
-import { Box, Typography, useTheme } from '@mui/material';
 import { TableColumnType } from '@terraware/web-components';
 import { useDeviceInfo } from '@terraware/web-components/utils';
 
@@ -9,8 +8,6 @@ import Page from 'src/components/Page';
 import Button from 'src/components/common/button/Button';
 import { APP_PATHS } from 'src/constants';
 import { useLocalization, useUser } from 'src/providers';
-import { selectCohorts } from 'src/redux/features/cohorts/cohortsSelectors';
-import { useAppSelector } from 'src/redux/store';
 import strings from 'src/strings';
 
 import CohortsTable from './CohortsTable';
@@ -37,7 +34,6 @@ const CohortListView = () => {
   const history = useHistory();
   const { isAllowed } = useUser();
   const canCreateCohorts = isAllowed('CREATE_COHORTS');
-  const cohorts = useAppSelector(selectCohorts);
 
   const goToNewCohort = () => {
     const newProjectLocation = {
@@ -69,46 +65,9 @@ const CohortListView = () => {
       }
       contentStyle={{ display: 'flex', flexGrow: 1, flexDirection: 'column' }}
     >
-      {!cohorts?.length ? (
-        <EmptyState onClick={goToNewCohort} />
-      ) : (
-        <CohortsTable columns={() => columns(activeLocale)} />
-      )}
+      <CohortsTable columns={() => columns(activeLocale)} />
     </Page>
   );
 };
 
 export default CohortListView;
-
-const EmptyState = ({ onClick }: { onClick: () => void }): JSX.Element => {
-  const theme = useTheme();
-  const { isAllowed } = useUser();
-
-  return (
-    <Box
-      sx={{
-        display: 'flex',
-        flexDirection: 'column',
-        flexGrow: 0,
-        margin: 'auto',
-        padding: theme.spacing(3, 3, 8),
-        textAlign: 'center',
-      }}
-    >
-      <Typography
-        color={theme.palette.TwClrTxt}
-        fontSize='16px'
-        fontWeight={400}
-        lineHeight='24px'
-        marginBottom={theme.spacing(2)}
-      >
-        {strings.COHORTS_EMPTY_STATE}
-      </Typography>
-      {isAllowed('CREATE_COHORTS') && (
-        <Box sx={{ margin: 'auto' }}>
-          <Button icon='plus' id='new-cohort' label={strings.ADD_COHORT} onClick={onClick} size='medium' />
-        </Box>
-      )}
-    </Box>
-  );
-};
